fix + notifica struttura senza convenzione

+++ plugins/Aziende/src/Model/Table/AgreementsToSediTable.php
@@ -67,7 +67,7 @@ class AgreementsToSediTable extends AppTable
             ->allowEmptyString('id', null, 'create');
 
         $validator
-            ->decimal('capacity')
+            ->integer('capacity')
             ->notEmptyString('capacity');
 
         $validator
